Addition methods, GUI actualized, classes actualized.
We implemented the methods to add an object from "RestaurantManagerGUI" to the data of "RestaurantManager" with dependency only between "RestaurantManagerGUI" and "RestaurantManager". Also actualized the constructor of some classes to take the id as a long , and not as a int. Finally, the window to create an user has been actualized with two more text area to put the name and last name of the employee who owns the new user.

+++ jfx-restaurantManager/src/model/Costumer.java
@@ -7,7 +7,7 @@ public class Costumer {
 	String sep=","; //separator
 	
 	//Optional.
-	int id;
+	long id;
 	
 	//Constructor whitout id.
 	public Costumer(String name, String lastname, String address, String observations, long phone) {
@@ -19,7 +19,7 @@ public class Costumer {
 	}
 	
 	//Constructor with id.
-	public Costumer(String name, String lastname, String address, String observations, long phone, int id) {
+	public Costumer(String name, String lastname, String address, String observations, long phone, long id) {
 		this.name = name;
 		this.lastname = lastname;
 		this.address = address;
