+++ jfx-restaurantManager/src/model/User.java
@@ -7,10 +7,10 @@ public class User extends Employee{
 	
 	//Employee variables.
 	String name, lastname;
-	int id;
+	long id;
 	String sep = ","; //separator
 	
-	public User(String userName, String password, String name, String lastname, int id) {
+	public User(String userName, String password, String name, String lastname, long id) {
 		super(name, lastname, id);
 		this.userName = userName;
 		this.password = password;
